Add menu with sound toggle & relaxation options
I'm planning to add two other types of relaxation:
a neverending train and some rain pouring down. Needs a bit more work
for switching the types and for playing the sounds belonging to those
new types.

Oh, and looping a sound is not working flawlessly yet. Need to look at
freesound.org on how to do that.